move Home components into Home folder
